- fixed StaticWeightedTotalScoreBetterThan compile error on gcc - fixed undefined DBL_MAX on gcc - fixed SQLite to link to libdl on gcc - updated freicore external
Former-commit-id: d0f8cfe6797c6bd5e8d335fa8b1bc0c45cf22e12

+++ pwiz_tools/Bumbershoot/idpicker/Qonverter/Parser.cpp
@@ -69,9 +69,9 @@ struct SharedStringFastLessThan
 
 struct AminoAcidTranslator
 {
-    static size_t size() {return 26;}
-    static size_t translate(char aa) {return aa - 'A';};
-    static char translate(size_t index) {return static_cast<char>(index) + 'A';}
+    static int size() {return 26;}
+    static int translate(char aa) {return aa - 'A';};
+    static char translate(int index) {return static_cast<char>(index) + 'A';}
 };
 
 struct IsNotAnalysisParameter
